Update useTooltips interface for consistency with usePopover
