+++ www/src/documentation/components/overlays/tooltip.mdx
@@ -82,6 +82,7 @@ If you need to be very specific
   {(props) => (
     <Button
       aria-describedby={props['aria-describedby']}
+      className={props.className}
       onBlur={props.onBlur}
       onFocus={props.onFocus}
       onMouseOut={props.onMouseOut}
